requirements.txt
pyvista
solidpython